Modified the external hook for Godmode/1HitKO feature, creating a pointer to current location of memory to easily calculate relative offsets needed for jumps.
Added 'preAddrForJne'

Modified the 'prepAddrForX' functions to use the current address instead of source address. The difference is only the name of the variable but now we are passing a memory address from a pointer that keeps tracking of our current location in memory while patching memory for a hook. The name change just helps remind me we aren't using the inital source address of the hook but the current address we should be after patching the previous opcodes into memory.