Added pointers and separated code
Added a raw pointer to a struct containing all the character offsets inside of the Class Character for an example of how to us raw pointers in a class correctly.

Create an object of ProcInfo procInfo upon the end of its declaration removing the need to call it before using it.

Changed the Menu menu object to a smart pointer "uniqueptr" to show an example of using smart pointers for objects.